Causal classifier (#8)
Adds a "causal" version of the classifier RM. Causal means that the
classifier RM uses the same architecture as the causal base model,
devoting the EOS token's logit to be the classifier logit.

Note that [this PR](https://github.com/databricks/compose-rl/pull/6)
introduces causal preference RMs. The current PR is based off of that
implementation, so it may make sense to wait on that PR or have this PR
merge into that branch.

This has been used extensively.
Here is an
[mlflow](https://dbc-559ffd80-2bfc.cloud.databricks.com/ml/experiments/4128924280951945/runs/d9f3526199b9402c8275786fb5b0ec14/model-metrics?o=7395834863327820)
run using the latest code. Notice how it runs.

---------

Co-authored-by: Brandon <bcui8377@gmail.com>
Co-authored-by: bcui-db <141345999+bcui-db@users.noreply.github.com>