Adding Token Counter for Online RL (#110)
We are updating the token counter to be defined by the `action_mask` +
`prompt_len`. We are also using the `len(sequences)` to count the number
of pad tokens for logging. Example runs: `compose-rl-grpo-test-WHoytM`

Why this change is needed:

The bug that we encounter comes from the fact that we use a
StreamingDataLoader without a bespoke `get_num_tokens_in_batch` fn (like
the one llmfoundry uses
[here](https://github.com/mosaicml/llm-foundry/blob/6048d9828e67a040959fba0082b84eeeaed7676a/llmfoundry/data/utils.py#L134)).
Therefore, we are using the default token counting function in Composer
[here](https://github.com/mosaicml/composer/blob/ef6315f5863dcfcdf9bd484e5d4d62cdfaae46de/composer/core/data_spec.py#L297).
As such, since we don't have `input_ids` in the batch, we end up just
using `max_seq_len * num_samples_in_batch` to count the total number of
tokens, which leads to the incorrect value that we're seeing in our
tests.