Fix for incorrect sequence of prompts being sent over for inference (#147)
This fixes the issue where the sequence of prompts we are sending over
to vLLM is incorrect and is actually causing a scattering pattern
through out its sequence wrt to the unique prompts generations where
what we want is that the sequence should have the generations be grouped
together for each unique prompt. This issue down the line will cause
cases where certain set ups would cause the model to not train as well
as this should fix the calculation for its rewards and advantage
calculation.

Runs with the fix:

Single node run:
https://dbc-559ffd80-2bfc.cloud.databricks.com/ml/experiments/3786561063037191/runs/6921b098214b4b0c88397c5ce0b170a2?o=7395834863327820
2 nodes run:
https://dbc-559ffd80-2bfc.cloud.databricks.com/ml/experiments/3786561063037191/runs/a9a74b1a4b2a4561941513d37c305550?o=7395834863327820